requirements.txt
biopython
joblib
matplotlib
pandas==1.4.1
seaborn
xlrd==1.2.0
distro
